requirements.txt
feedparser==5.2.1
requests==2.22.0
beautifulsoup4==4.8.2
pandas==1.0.3
newspaper3k==0.2.8
nltk==3.4.5

